chore: update button label to 'Simpan Guru / Staff' and improve photo validation logic

+++ app/Livewire/Dashboard/GuruStaff/Edit.php
@@ -39,8 +39,8 @@ class Edit extends Component
     {
         $data = Collection::make($this->validate());
 
-        if (! is_null($data->get('photo'))) {
-            if (Storage::exists($this->employee->photo)) {
+        if (! blank($data->get('photo'))) {
+            if ($this->employee->photo && Storage::exists($this->employee->photo)) {
                 Storage::delete($this->employee->photo);
             }
